Slight changes for the data to be more fitted for the API call and in the comments is  API call itself!

+++ src/TrashFound.h
@@ -3,7 +3,7 @@
 /// @brief Structure to hold information about found trash
 struct trash{
     int type; //< Trash classification
-    int confidence; //< Confidence level of the classification
+    float confidence; //< Confidence level of the classification
     float latitude; //< Latitude of the trash location
     float longitude; //< Longitude of the trash location
 };
